More battle system

+++ bot/commands/battle.py
@@ -42,6 +42,7 @@ async def challenge(ctx):
 
   session.commit()
 
+
 # On message received
 @client.event
 async def on_message(message):
@@ -54,38 +55,41 @@ async def on_message(message):
     session.add(user)
   p2_battle = user.current_battle_p2
 
-  if p2_battle != None:
+  if p2_battle != None and p2_battle.accepted == False:
     if message.content.lower() in ['y', 'yes']:
       p2_battle.accepted = True
       await getdiscorduser(p2_battle.player1.handle).send(f"{message.author.name} has accepted your challenge!\n\nWhat would you like to do?\n1) Attack")
 
     elif message.content.lower() in ['n', 'no']:
       session.delete(p2_battle)
-    else:
-      await message.channel.send("Sorry, didn't understand you")
+  else:
+    p1_battle = user.current_battle_p1
+    battle = None
 
-  p1_battle = user.current_battle_p1
-  battle = None
+    if p1_battle is not None and p1_battle.is_my_turn(user):
+      battle = p1_battle
+    if p2_battle is not None and p2_battle.is_my_turn(user):
+      battle = p2_battle
 
-  if p1_battle is not None and p1_battle.is_my_turn(user):
-    battle = p1_battle
-  if p2_battle is not None and p2_battle.is_my_turn(user):
-    battle = p2_battle
-
-  if battle is not None:
-    if message.content.lower() in ['1']:
-      still_going = p1_battle.do_turn(1)
-      attacker = getdiscorduser(battle.attackerPlayer().handle)
-      defender = getdiscorduser(battle.defenderPlayer().handle)
-      if not still_going:
-        await attacker.send(f"Aw yiss! You get all the breadcrumbs.")
-        await defender.send(f"You lost. Damn")
+    if battle is not None:
+      if message.content.lower() in ['1']:
+        result = p1_battle.do_turn(1)
+        attacker = getdiscorduser(battle.attackerPlayer().handle)
+        attack_duck_name = battle.defender().nickname
+        defender = getdiscorduser(battle.defenderPlayer().handle)
+        if result == False:
+          xp, level_up = battle.win(attacker)
+          await attacker.send(f"Aw yiss! You get all the breadcrumbs.\n{attack_duck_name} ate up {xp} breadcrumbs. Nom nom nom")
+          if level_up > 0:
+            await attacker.send(f"Holey guacamoley! {attack_duck_name} leveled up!")
+          await defender.send(f"You lost. Damn")
+        else:
+          damage, remaining_hp = result
+          defend_duck_name = battle.defender().nickname
+          await attacker.send(f"{attack_duck_name} attacked for {damage} points! {defend_duck_name} has {remaining_hp} HP left.\nWaiting for your opponent to make a move...")
+          await defender.send(f"{defend_duck_name} got walloped and has {remaining_hp} HP left.\n\nWhat would you like to do?\n1) Attack")
       else:
-        duck_name = battle.defender().nickname
-        remaining_hp = battle.defender().current_hp
-        await defender.send(f"{duck_name} got walloped and has {remaining_hp} HP left.\n\nWhat would you like to do?\n1) Attack")
-    else:
-      await message.channel.send("Sorry, didn't understand you")
+        await message.channel.send("Sorry, didn't understand you")
 
   session.commit()
   await client.process_commands(message)
